Converts AppDefaults to singleton
