+++ Shared/Extensions/AddWebFeedDefaultContainer.swift
@@ -13,8 +13,8 @@ struct AddWebFeedDefaultContainer {
 	
 	static var defaultContainer: Container? {
 		
-		if let accountID = AppDefaults.addWebFeedAccountID, let account = AccountManager.shared.activeAccounts.first(where: { $0.accountID == accountID }) {
-			if let folderName = AppDefaults.addWebFeedFolderName, let folder = account.existingFolder(withDisplayName: folderName) {
+		if let accountID = AppDefaults.shared.addWebFeedAccountID, let account = AccountManager.shared.activeAccounts.first(where: { $0.accountID == accountID }) {
+			if let folderName = AppDefaults.shared.addWebFeedFolderName, let folder = account.existingFolder(withDisplayName: folderName) {
 				return folder
 			} else {
 				return substituteContainerIfNeeded(account: account)
@@ -28,11 +28,11 @@ struct AddWebFeedDefaultContainer {
 	}
 	
 	static func saveDefaultContainer(_ container: Container) {
-		AppDefaults.addWebFeedAccountID = container.account?.accountID
+		AppDefaults.shared.addWebFeedAccountID = container.account?.accountID
 		if let folder = container as? Folder {
-			AppDefaults.addWebFeedFolderName = folder.nameForDisplay
+			AppDefaults.shared.addWebFeedFolderName = folder.nameForDisplay
 		} else {
-			AppDefaults.addWebFeedFolderName = nil
+			AppDefaults.shared.addWebFeedFolderName = nil
 		}
 	}
