+++ Shared/Extensions/CacheCleaner.swift
@@ -15,8 +15,8 @@ struct CacheCleaner {
 
 	static func purgeIfNecessary() {
 
-		guard let flushDate = AppDefaults.lastImageCacheFlushDate else {
-			AppDefaults.lastImageCacheFlushDate = Date()
+		guard let flushDate = AppDefaults.shared.lastImageCacheFlushDate else {
+			AppDefaults.shared.lastImageCacheFlushDate = Date()
 			return
 		}
 		
@@ -41,7 +41,7 @@ struct CacheCleaner {
 						}
 					}
 					
-					AppDefaults.lastImageCacheFlushDate = Date()
+					AppDefaults.shared.lastImageCacheFlushDate = Date()
 					
 				}
 			}
